refactor(cli): split output-formatting and parser construction into services
cli.py had already gone through one service-module split (docker, key,
package, workflow, logs, projects) in v2.28.0. What remained was
argparse plumbing, dispatch, and output printing -- two more clean,
self-contained cuts:

- services/cli_output.py: _print_payload, _format_size, _queue_state,
  _print_status, _print_doctor, _print_maintenance. Pure functions,
  zero coupling to argparse or dispatch.
- services/cli_parser.py: _product_parser, _compatibility_parser,
  _add_profile_arguments, _add_docker_run_arguments,
  _ProductArgumentParser, _product_help. Pure argparse construction,
  no runtime side effects.

_dispatch_product, main(), _run_foreground, and the runtime-preset/
dispatch helpers stay in cli.py -- that's the real orchestration and
belongs in the owner file. cli.py: 786 -> 509 lines.

Verbatim extraction, no behavior change. Opus-reviewed; the review
caught one real miss (tests/test_runtime_cli.py:317 still called
cli._product_help() after it was correctly dropped from cli.py's
imports) -- fixed by importing from the new module directly, matching
the fix already applied in test_command_smoke.py. Also added direct
unit tests for _format_size/_queue_state (services/cli_output.py had
zero dedicated coverage before, only indirect exercise through
existing cli.py tests).

Full suite: 810 passed, 29 skipped, same 6 pre-existing sandbox-only
failures (bundled spaCy concept model unavailable in this environment,
confirmed present on a clean MARM-main checkout before this change).
